ECDSAStakeRegistry is inherited in BitDSMRegistry
ECDSAStakeRegistry.sol is modified as abstract contract to act as base for BitDSMRegistry contract. This is done to streamline the process of operator registration and deregistration. BitDSMRegistry is the entry point for managing operator registration with BitcoinPodManager and the AVS service. AVS service shall now be initialized with BitDSMRegistry contract address instead of StakeREgistry address.